+++ .circleci/config.yml
@@ -31,24 +31,14 @@ jobs:
             pytest tests/test_materials.py -v --cov=openmc_dagmc_wrapper --cov-append --cov-report term --cov-report xml
 
       - run:
-          name: run tests MeshTally()
+          name: run tests Tallies
           command: 
-            pytest tests/test_tallies/test_mesh_tallies.py -v --cov=openmc_dagmc_wrapper --cov-append --cov-report term --cov-report xml
+            pytest tests/test_tallies/ -v --cov=openmc_dagmc_wrapper --cov-append --cov-report term --cov-report xml
 
       - run:
           name: System tests
           command: 
-            pytest tests/test_system.py -v --cov=openmc_dagmc_wrapper --cov-append --cov-report term --cov-report xml
-
-      - run:
-          name: run test_shape_neutronics
-          command: 
-            pytest tests/test_shape_neutronics.py -v --cov=openmc_dagmc_wrapper --cov-append --cov-report term --cov-report xml
-
-      - run:
-          name: run test_reactor_neutronics
-          command: 
-            pytest tests/test_reactor_neutronics.py -v --cov=openmc_dagmc_wrapper --cov-append --cov-report term --cov-report xml
+            pytest tests/test_system/ -v --cov=openmc_dagmc_wrapper --cov-append --cov-report term --cov-report xml
 
       # - run:
       #     name: run test_example_neutronics_simulations
