+++ tests/test_system/test_system_multi_volume.py
@@ -1,4 +1,3 @@
-
 import os
 import tarfile
 import unittest
@@ -76,36 +75,6 @@ class TestObjectNeutronicsArguments(unittest.TestCase):
         results = odw.process_results(statepoint_filename=h5m_filename)
 
         assert results["TBR"]["result"] > 0.0
-
-    def test_bounding_box_size(self):
-
-        bounding_box = odw.find_bounding_box(self.h5m_filename_bigger)
-
-        print(bounding_box)
-        assert len(bounding_box) == 2
-        assert len(bounding_box[0]) == 3
-        assert len(bounding_box[1]) == 3
-        assert bounding_box[0][0] == pytest.approx(-10005, abs=0.1)
-        assert bounding_box[0][1] == pytest.approx(-10005, abs=0.1)
-        assert bounding_box[0][2] == pytest.approx(-10005, abs=0.1)
-        assert bounding_box[1][0] == pytest.approx(10005, abs=0.1)
-        assert bounding_box[1][1] == pytest.approx(10005, abs=0.1)
-        assert bounding_box[1][2] == pytest.approx(10005, abs=0.1)
-
-    def test_bounding_box_size_2(self):
-
-        bounding_box = odw.find_bounding_box(self.h5m_filename_smaller)
-
-        print(bounding_box)
-        assert len(bounding_box) == 2
-        assert len(bounding_box[0]) == 3
-        assert len(bounding_box[1]) == 3
-        assert bounding_box[0][0] == pytest.approx(-10005, abs=0.1)
-        assert bounding_box[0][1] == pytest.approx(-10005, abs=0.1)
-        assert bounding_box[0][2] == pytest.approx(-10005, abs=0.1)
-        assert bounding_box[1][0] == pytest.approx(10005, abs=0.1)
-        assert bounding_box[1][1] == pytest.approx(10005, abs=0.1)
-        assert bounding_box[1][2] == pytest.approx(10005, abs=0.1)
 
 
 # # move to neutronics_workflow
